fix: address Sentry review feedback
