+++ openspec/changes/add-sentry-sdk/specs/sentry-observability/spec.md
@@ -14,6 +14,10 @@ The application SHALL load `WGA_SENTRY_DSN` through `internal/config` as optiona
 - **WHEN** `WGA_SENTRY_DSN` is unset or empty
 - **THEN** the application starts normally and emits one structured log entry that Sentry monitoring is disabled
 
+#### Scenario: DSN is malformed
+- **WHEN** `WGA_SENTRY_DSN` is supplied in an invalid format
+- **THEN** server configuration fails with an operator-safe validation error that identifies `WGA_SENTRY_DSN`
+
 ### Requirement: Server error monitoring
 
 The serving runtime SHALL initialise the official Sentry Go SDK when a valid DSN is configured, SHALL label events with the configured WGA environment, and SHALL use bounded flushing during shutdown. It SHALL capture unhandled server failures without changing the response or error-propagation behaviour observed by clients.
@@ -46,6 +50,13 @@ The browser bundle SHALL include the official Sentry browser SDK and SHALL initi
 
 - **WHEN** a full page is rendered with Sentry monitoring disabled
 - **THEN** the browser skips Sentry initialisation and the main application bootstrap still completes
+
+### Requirement: Browser event URL privacy
+The browser SDK SHALL remove query strings and fragments from error-event and breadcrumb URLs before sending events to Sentry.
+
+#### Scenario: Page URL contains a sensitive query parameter
+- **WHEN** a browser error occurs on a URL containing query parameters or a fragment
+- **THEN** the event sent to Sentry contains the URL path without the query string or fragment
 
 ### Requirement: Public configuration boundary
