+++ resources/js/sentry.ts
@@ -5,6 +5,38 @@ export type SentryConfiguration = {
 	environment: string;
 };
 
+const scrubURL = (value: string): string => {
+	try {
+		const isRelative = value.startsWith("/");
+		const url = new URL(value, "https://wga.invalid");
+		url.search = "";
+		url.hash = "";
+		if (isRelative) {
+			return url.pathname;
+		}
+		return url.toString();
+	} catch {
+		return value;
+	}
+};
+
+export const scrubSentryEvent = (event: Sentry.Event): Sentry.Event => ({
+	...event,
+	request: event.request?.url
+		? { ...event.request, url: scrubURL(event.request.url) }
+		: event.request,
+	breadcrumbs: event.breadcrumbs?.map((breadcrumb) => {
+		if (typeof breadcrumb.data?.url !== "string") {
+			return breadcrumb;
+		}
+
+		return {
+			...breadcrumb,
+			data: { ...breadcrumb.data, url: scrubURL(breadcrumb.data.url) },
+		};
+	}),
+});
+
 type MetadataDocument = {
 	querySelector: (selector: string) => {
 		getAttribute: (name: string) => string | null;
@@ -32,9 +64,19 @@ export const initialiseSentry = (
 		return false;
 	}
 
-	initialise({
-		dsn: configuration.dsn,
-		environment: configuration.environment,
-	});
-	return true;
+	try {
+		const client = initialise({
+			dsn: configuration.dsn,
+			environment: configuration.environment,
+			beforeSend: scrubSentryEvent,
+		});
+		if (!client) {
+			console.error("Sentry browser initialisation failed");
+			return false;
+		}
+		return true;
+	} catch (error) {
+		console.error("Sentry browser initialisation failed", error);
+		return false;
+	}
 };
